Update to 4.4.0.55262

+++ SpellWork/DBC/Structures/DifficultyEntry.cs
@@ -13,7 +13,7 @@ namespace SpellWork.DBC.Structures
         public byte FallbackDifficultyID;
         public byte MinPlayers;
         public byte MaxPlayers;
-        public ushort Flags;
+        public byte Flags;
         public byte ItemContext;
         public byte ToggleDifficultyID;
         public ushort GroupSizeHealthCurveID;
