+++ SpellWork/DBC/Structures/SpellAuraOptionsEntry.cs
@@ -7,13 +7,13 @@ namespace SpellWork.DBC.Structures
         [Index(true)]
         public uint ID;
         public byte DifficultyID;
-        public ushort CumulativeAura;
+        public uint CumulativeAura;
         public int ProcCategoryRecovery;
         public byte ProcChance;
         public int ProcCharges;
         public ushort SpellProcsPerMinuteID;
         [Cardinality(2)]
         public int[] ProcTypeMask = new int[2];
-        public int SpellID;
+        public uint SpellID;
     }
 }
