+++ SpellWork/DBC/Structures/SpellXSpellVisualEntry.cs
@@ -9,7 +9,7 @@ namespace SpellWork.DBC.Structures
         public byte DifficultyID;
         public uint SpellVisualID;
         public float Probability;
-        public int Flags;
+        public byte Flags;
         public int Priority;
         public int SpellIconFileID;
         public int ActiveIconFileID;
